requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
lxml==4.8.0
pydantic==1.9.0
python-decouple==3.6
python-dotenv==0.20.0
requests==2.27.1
soupsieve==2.3.1
typing-extensions==4.2.0
urllib3==1.26.9
